pyproject.toml
[tool.poetry]
name = "tts-bot"
version = "0.0.1"
description = "tts bot"
authors = ["MisileLaboratory"]
license = "MisileLab License"

[tool.poetry.dependencies]
python = "^3.10"
disnake = {extras = ["voice"], version = "^2.2.2"}
google-cloud-texttospeech = "^2.9.0"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
pylint = "^2.12.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
